added the ConfirmUser route
