+++ app/controllers/api/v1/confirmations_controller.rb
@@ -1,25 +1,7 @@
 class Api::V1::ConfirmationsController < Devise::ConfirmationsController
 
-  def create
-    
-    @user = User.find_by(id: params[:user_id])
-    render json: @user
-  
-    # if @user
-    #   jwt = JWT.encode(
-    #     {
-    #       user_id: @user.id, # the data to encode
-    #       exp: 24.hours.from_now.to_i # the expiration time
-    #     },
-    #     Rails.application.credentials.fetch(:secret_key_base), # the secret key
-    #     'HS256' # the encryption algorithm
-    #   )
-    #   render json: {jwt: jwt, email: @user.email, user_id: @user.id}, status: :created
-    # else
-    #   render json: {}, status: :unauthorized
-    # end
+
+
 
   
-  end 
-
 end
