+++ frontend/src/components/SignIn/SignIn.js
@@ -28,6 +28,7 @@ const SignIn = ({ history, setJwt }) => {
       .catch((error) => {
         setError(error.response.data.errors[0].message);
       });
+      
   };
 
   return (
